dot: refactor label formatting
